+++ R/ipcw_logistic_regression.R
@@ -2,17 +2,19 @@
 #'
 #' @description
 #' \loadmathjax
-#' Fits a logistic regression model with IPCW for
-#' right-censored survival data. Jackknife model training is performed.
+#' Fits an IPCW-weighted logistic regression for right-censored
+#' survival data.
 #'
 #' @details
-#' Training is performed using the `glm`
-#' using a quasibinomial family to account for the weights.
+#' A logistic model is fitted to the full dataset. Jackknife refits are
+#' computed to derive jackknife-based standard errors.
+#' Training is performed using `glm` with a binomial distribution to
+#' account for the IPCW weights.
 #'
 #' @inheritParams ipcw_xgboost
 #' @inherit ipcw_xgboost return
 #' @seealso [ipcw_weights()] for the underlying implementation of the weights
-#' and [IPCWJK] for more information.
+#' and [IPCWJK] as well as \insertCite{paper}{IPCWJK} for more information.
 #' @importFrom stats glm as.formula predict
 #' @importFrom Rdpack reprompt
 #' @import mathjaxr
@@ -30,7 +32,8 @@
 #' predict(fit, newdata)
 #' @export
 ipcw_logistic_regression <- function(
-    data, tau, time_var = "t", status_var = "delta") {
+  data, tau, time_var = "t", status_var = "delta"
+) {
   n <- nrow(data)
   w <- ipcw_weights(data, tau, time_var = time_var, status_var = status_var)
   w <- w / sum(w)
@@ -40,10 +43,21 @@ ipcw_logistic_regression <- function(
   dtrain <- as.data.frame(cbind(data[, training_vars], y = y))
   colnames(dtrain) <- c(training_vars, "y")
   formula <- as.formula(paste("y ~", paste(training_vars, collapse = "+")))
-  full_model <- glm(formula,
-    data = dtrain, weights = w,
-    family = "quasibinomial"
+  full_model <- NULL
+  withCallingHandlers(
+    {
+      full_model <- glm(formula,
+        data = dtrain, weights = w,
+        family = "binomial"
+      )
+    },
+    warning = function(w) {
+      if (grepl("non-integer", conditionMessage(w))) {
+        invokeRestart("muffleWarning") # suppress this one
+      }
+    }
   )
+
 
   train_pred <- predict(full_model, dtrain, type = "response")
   train_brier <- sum(w * (train_pred - y)**2)
@@ -53,10 +67,21 @@ ipcw_logistic_regression <- function(
     wadj <- (n - 1) * w[-i] / (sum(w[-i]))
     dtraini <- as.data.frame(cbind(data[-i, training_vars], y = y[-i]))
     colnames(dtraini) <- c(training_vars, "y")
-    model_wjk <- glm(formula,
-      data = dtraini, weights = wadj,
-      family = "quasibinomial"
+    model_wjk <- NULL
+    withCallingHandlers(
+      {
+        model_wjk <- glm(formula,
+          data = dtraini, weights = wadj,
+          family = "binomial"
+        )
+      },
+      warning = function(w) {
+        if (grepl("non-integer", conditionMessage(w))) {
+          invokeRestart("muffleWarning") # suppress this one
+        }
+      }
     )
+
     jk_models[[i]] <- model_wjk
   }
   predfun <- function(model, x) {