+++ NEWS.md
@@ -1,6 +1,24 @@
 # IPCWJK (development version)
 
-* Initial package functionality
+* Suppressed the warning about the non-integer successes in the binomial model
+
+## Minor improvements and bug fixes
+
+* Added references to the paper
+* Improved documentation
+* Closes #17, #16, #15, #14, #13, #10
+
+# IPCWJK 0.1.1.9000
+
+## Major changes
+
+* Suppressed the warning about the non-integer successes in the binomial model
+
+## Minor improvements and bug fixes
+
+* Added references to the paper
+* Improved documentation
+* Closes #17, #16, #15, #14, #13, #10
 
 # IPCWJK 0.1.0.9000